+++ src/pages/layout/header/Header.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import { headerContainer, headerLeftContainer, headerLinkContainer, headerMainContainer, headerMainIconContainer, headerProfileContainer, headerRightContainer, headerSocialContainer } from './style';
-import alertIcon from './img/alert.png';
-import messageIcon from './img/message.png';
-import persenalBuddyIcon from './img/persenalBuddyIcon.png';
-import memberProfile from './img/memberProfile.png';
+import alertIcon from '../../../../public/assets/images/header/alert.png';
+import messageIcon from '../../../../public/assets/images/header/message.png';
+import persenalBuddyIcon from '../../../../public/assets/images/header/persenalBuddyIcon.png';
+import memberProfile from '../../../../public/assets/images/header/memberProfile.png';
 
 
 const Header = () => {
